.history/backend/requirements_20250908142626.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud Services
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4
google-cloud-aiplatform==1.38.1
firebase-admin==6.4.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Email APIs (Gmail/Office365)
msal==1.25.0

# AI/ML Processing
google-generativeai==0.3.2
vertexai==1.38.1
torch==2.1.1
torchvision==0.16.1
transformers==4.36.2

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3

# HTTP & API
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Environment & Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Deployment
gunicorn==21.2.0